test(domain): Lotto 생성 시 정상 케이스 테스트 추가
- 6개의 번호로 로또 생성 테스트
- @Nested로 생성 관련 테스트 그룹화
- 기존 제공 테스트 유지
